Separated the tests and project code into different packages. This should help with test discovery.
