+++ from_pattern_draft/pattern_draft_to_output.py
