+++ tests/test_unit_tests_for_pattern_draft_to_output.py
@@ -1,4 +1,4 @@
-import pattern_draft_to_output
+from from_pattern_draft import pattern_draft_to_output
 import unittest
 
 
@@ -25,53 +25,53 @@ class TestCardInstantiation(unittest.TestCase):
     def test_raises_TypeError_with_not_int_for_number_of_holes (self):
         self.assertRaises(TypeError,
                           pattern_draft_to_output.Card, number_of_holes='4',
-                                                        threading_direction='LTR',
-                                                        list_of_thread_colours=self.expected_thread_colours,
-                                                        list_of_thread_types=self.expected_thread_types
+                          threading_direction='LTR',
+                          list_of_thread_colours=self.expected_thread_colours,
+                          list_of_thread_types=self.expected_thread_types
                           )
 
     def test_raises_TypeError_with_not_str_for_threading_direction(self):
         self.assertRaises(TypeError,
                           pattern_draft_to_output.Card, number_of_holes=self.expected_number_of_holes,
-                                                        threading_direction=['LTR'],
-                                                        list_of_thread_colours=self.expected_thread_colours,
-                                                        list_of_thread_types=self.expected_thread_types
+                          threading_direction=['LTR'],
+                          list_of_thread_colours=self.expected_thread_colours,
+                          list_of_thread_types=self.expected_thread_types
                           )
 
     def test_raises_ValueError_with_incorrect_str_for_threading_direction(self):
         self.assertRaises(ValueError,
                           pattern_draft_to_output.Card, number_of_holes=self.expected_number_of_holes,
-                                                        threading_direction='AAA',
-                                                        list_of_thread_colours=self.expected_thread_colours,
-                                                        list_of_thread_types=self.expected_thread_types
+                          threading_direction='AAA',
+                          list_of_thread_colours=self.expected_thread_colours,
+                          list_of_thread_types=self.expected_thread_types
                           )
 
     def test_raises_ValueError_with_incorrect_length_for_list_of_thread_colours(self):
         self.assertRaises(ValueError,
                           pattern_draft_to_output.Card, number_of_holes=self.expected_number_of_holes,
-                                                        threading_direction='LTR',
-                                                        list_of_thread_colours=['RRR', 'BBB', 'YYY'],
-                                                        list_of_thread_types=self.expected_thread_types
+                          threading_direction='LTR',
+                          list_of_thread_colours=['RRR', 'BBB', 'YYY'],
+                          list_of_thread_types=self.expected_thread_types
                           )
         self.assertRaises(ValueError,
                           pattern_draft_to_output.Card, number_of_holes=self.expected_number_of_holes,
-                                                        threading_direction='LTR',
-                                                        list_of_thread_colours=['RRR', 'BBB', 'YYY', 'WWW', 'XXX'],
-                                                        list_of_thread_types=self.expected_thread_types
+                          threading_direction='LTR',
+                          list_of_thread_colours=['RRR', 'BBB', 'YYY', 'WWW', 'XXX'],
+                          list_of_thread_types=self.expected_thread_types
                           )
 
     def test_raises_ValueError_with_incorrect_length_for_list_of_thread_types(self):
         self.assertRaises(ValueError,
                           pattern_draft_to_output.Card, number_of_holes=self.expected_number_of_holes,
-                                                        threading_direction='LTR',
-                                                        list_of_thread_colours=self.expected_thread_colours,
-                                                        list_of_thread_types=['CCC', 'CCC', 'CCC']
+                          threading_direction='LTR',
+                          list_of_thread_colours=self.expected_thread_colours,
+                          list_of_thread_types=['CCC', 'CCC', 'CCC']
                           )
         self.assertRaises(ValueError,
                           pattern_draft_to_output.Card, number_of_holes=self.expected_number_of_holes,
-                                                        threading_direction='LTR',
-                                                        list_of_thread_colours=self.expected_thread_colours,
-                                                        list_of_thread_types=['CCC', 'CCC', 'CCC', 'CCC', 'XXX']
+                          threading_direction='LTR',
+                          list_of_thread_colours=self.expected_thread_colours,
+                          list_of_thread_types=['CCC', 'CCC', 'CCC', 'CCC', 'XXX']
                           )
 
 
@@ -212,3 +212,7 @@ class TestStackFunctions(unittest.TestCase):
         self.assertEqual(top_typs, ['cotton', 'cotton'])
         self.assertEqual(bottom_cols, ['Blue', 'B'])
         self.assertEqual(bottom_typs, ['cotton', 'cotton'])
+
+
+if __name__ == '__main__':
+    unittest.main()
